Revert "Set lmdb.NoReadahead flag conditionally"
This reverts commit e33adc49f3c7b0e5f917ced8a8706cd1844a5b52.

We had a discussion about this. It does not seem appropriate to
run the read and the iterate benchmarks under different
settings.

We do not do this for other key value stores in our benchmark.

Moreover, in a real-world scenario, one does not reopen a KV store to
just iterate on the keys. So it seems fair that the benchmark should
retain the same settings for random reads and iterates.

Reverting this commit means that iterates run with lmdb.NoReadahead
flag set, just like the random read benchmarks.